+++ ReleaseNotes.md
@@ -1,3 +1,15 @@
+# v4.NEXT
+
+## Summary
+
+## Upgrade Instructions
+
+## Features
+
+## Bug fixes
+- cta/CTA#1101 - Fix disk space reservation logic adding all existing disk space reservations for all disk systems
+- cta/CTA#1023 - Retrieve puts the queue to sleep if the eos disk instance is not reachable
+
 # v4.5.0-1
 
 ## Summary
@@ -15,7 +27,6 @@
 ## Bug fixes
 - cta/CTA#1092 - Fix overflow error with drive state latestBandwith causing cta frontend crash
 - cta/CTA#501 - Fix disappearing reason when TapeDrive is reading or writing
-- cta/CTA#1101 - Fix disk space reservation logic adding all existing disk space reservations for all disk systems
 
 # v4.4.1-1